building out database
